Refactoring delete account endpoint

+++ src/routes/accounts/handlers/delete-account.ts
@@ -1,11 +1,11 @@
 import { z } from 'zod'
 
-import { ServerRequest, ServerResponse } from '@/lib/types'
+import { ERROR_CODES, ServerError, ServerRequest, ServerResponse } from '@/lib/types'
 
 import prisma from '@/services/prisma'
 
 export const DeleteAccountSchema = z.object({
-  body: z.object({
+  params: z.object({
     accountId: z.string(),
   }),
 })
@@ -16,9 +16,13 @@ type Request = z.infer<typeof DeleteAccountSchema>
  * This removes connection between the user and account
  * and deletes the account if no other users are connected to it
  */
-export const deleteAccount = async (req: ServerRequest<Request['body']>, res: ServerResponse) => {
+export const deleteAccount = async (req: ServerRequest<object, Request['params']>, res: ServerResponse) => {
   const { userId } = res.locals
-  const { accountId } = req.body
+  const { accountId } = req.params
+
+  if (!userId) {
+    throw new ServerError(401, ERROR_CODES.UNAUTHORIZED)
+  }
 
   await prisma.accounts.delete({
     where: {
